[move] log and getit -> config dir
