+++ lib/blocs/auth/login/login_cubit.dart
@@ -1,8 +1,8 @@
 import 'package:bloc/bloc.dart';
 import 'package:dalal_street_client/blocs/dalal/dalal_bloc.dart';
+import 'package:dalal_street_client/config/log.dart';
 import 'package:dalal_street_client/constants/error_messages.dart';
 import 'package:dalal_street_client/grpc/client.dart';
-import 'package:dalal_street_client/main.dart';
 import 'package:dalal_street_client/models/company_info.dart';
 import 'package:dalal_street_client/pages/auth/login_page.dart';
 import 'package:dalal_street_client/global_streams.dart';
