+++ lib/blocs/auth/register/register_cubit.dart
@@ -1,7 +1,7 @@
 import 'package:bloc/bloc.dart';
+import 'package:dalal_street_client/config/log.dart';
 import 'package:dalal_street_client/constants/error_messages.dart';
 import 'package:dalal_street_client/grpc/client.dart';
-import 'package:dalal_street_client/main.dart';
 import 'package:dalal_street_client/proto_build/actions/Register.pb.dart';
 import 'package:equatable/equatable.dart';
