+++ lib/blocs/auth/verify_phone/enter_phone/enter_phone_cubit.dart
@@ -1,8 +1,9 @@
 import 'package:bloc/bloc.dart';
 import 'package:dalal_street_client/blocs/dalal/dalal_bloc.dart';
+import 'package:dalal_street_client/config/get_it.dart';
+import 'package:dalal_street_client/config/log.dart';
 import 'package:dalal_street_client/constants/error_messages.dart';
 import 'package:dalal_street_client/grpc/client.dart';
-import 'package:dalal_street_client/main.dart';
 import 'package:dalal_street_client/proto_build/actions/AddPhone.pb.dart';
 import 'package:equatable/equatable.dart';
