+++ lib/blocs/daily_challenges/single_day_challenges/single_day_challenges_cubit.dart
@@ -1,7 +1,8 @@
 import 'package:bloc/bloc.dart';
+import 'package:dalal_street_client/config/get_it.dart';
+import 'package:dalal_street_client/config/log.dart';
 import 'package:dalal_street_client/constants/error_messages.dart';
 import 'package:dalal_street_client/grpc/client.dart';
-import 'package:dalal_street_client/main.dart';
 import 'package:dalal_street_client/models/daily_challenge_info.dart';
 import 'package:dalal_street_client/proto_build/actions/GetDailyChallenges.pb.dart';
 import 'package:dalal_street_client/proto_build/actions/GetMyUserState.pb.dart';
