+++ lib/blocs/dalal/dalal_bloc.dart
@@ -1,5 +1,6 @@
+import 'package:dalal_street_client/config/get_it.dart';
+import 'package:dalal_street_client/config/log.dart';
 import 'package:dalal_street_client/grpc/client.dart';
-import 'package:dalal_street_client/main.dart';
 import 'package:dalal_street_client/models/company_info.dart';
 import 'package:dalal_street_client/global_streams.dart';
 import 'package:dalal_street_client/proto_build/actions/Login.pb.dart';
