+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:dalal_street_client/blocs/dalal/dalal_bloc.dart';
-import 'package:dalal_street_client/config.dart';
+import 'package:dalal_street_client/config/config.dart';
 import 'package:dalal_street_client/constants/error_messages.dart';
 import 'package:dalal_street_client/grpc/client.dart';
 import 'package:dalal_street_client/navigation/route_generator.dart';
@@ -8,52 +8,10 @@ import 'package:dalal_street_client/utils/snackbar.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:get_it/get_it.dart';
+import 'package:dalal_street_client/config/get_it.dart';
 import 'package:hydrated_bloc/hydrated_bloc.dart';
-import 'package:logger/logger.dart';
+import 'package:dalal_street_client/config/log.dart';
 import 'package:path_provider/path_provider.dart';
-
-/// Logging
-final Logger logger = Logger();
-
-/// ## Dependency Injection
-///
-/// ### Registering objects
-/// #### Register objects to then read them from anywhere in the project
-/// ```dart
-/// // Most of the time getIt can differentiate between objects from thier Type
-/// getIt.registerSingleton<String>(sessionId);
-/// // Or like this, as type can be inferred in dart
-/// getIt.registerSingleton(sessionId);
-///
-/// // But if you have multiple objects of same type, then mention a name
-/// getIt.registerSingleton(sessionId);
-/// getIt.registerSingleton(someOtherString, instanceName: 'String 2');
-/// ```
-///
-/// ### Reading objects
-/// #### Warning: Trying to read unregistered objects will throw exception
-/// ```dart
-/// final sessionId = getIt<String>();
-/// // or
-/// final String sessionId = getIt();
-/// // reading objects registered with a name
-/// final String string2 = getIt(instanceName: 'String 2');
-/// ```
-///
-/// ### Unregistering objects
-/// ```dart
-/// // Unregister objects with thier type
-/// getIt.unregister<String>();
-/// getIt.unregister<User>();
-/// // Can unregister with names
-/// getIt.unregister(instanceName: 'String 2');
-/// // Or unregister everything:
-/// getIt.reset();
-/// ```
-///
-/// For more complex usecases checkout [get_it](https://pub.dev/packages/get_it)
-final GetIt getIt = GetIt.I;
 
 void main() async {
   // Something doesnt work without this line. Dont remember what
