+++ lib/pages/daily_challenges/daily_challenge_item.dart
@@ -1,4 +1,4 @@
-import 'package:dalal_street_client/main.dart';
+import 'package:dalal_street_client/config/get_it.dart';
 import 'package:dalal_street_client/models/company_info.dart';
 import 'package:dalal_street_client/models/daily_challenge_info.dart';
 import 'package:dalal_street_client/proto_build/models/DailyChallenge.pb.dart';
